### 2017-12-12 #169 @cboettig
rdflib
`rdflib` is simply a wrapper around two existing ropensci packages: `redland` and `jsonld`, which should be a user-friendly complement to the low-level interface already provided by  `redland` for working with RDF (semantic/linked data).

```
Package: rdflib
Title: Tools to Manipulate and Query Semantic Data
Version: 0.0.2
Authors@R: person("Carl", "Boettiger",
                  email = "cboettig@gmail.com",
                  comment=c(ORCID = "http://orcid.org/0000-0002-1642-628X"))
Description: The Resource Description Framework, or 'RDF' is a widely used
             data representation model that forms the cornerstone of the
             Semantic Web. 'RDF' represents data as a graph rather than
             the familiar data table or rectangle of relational databases.
             The 'rdflib' package provides a friendly and concise user interface
             for performing common tasks on 'RDF' data, such as reading, writing
             and converting between the various serializations of 'RDF' data,
             including 'rdfxml', 'turtle', 'nquads', 'ntriples', 'trig', and 'json-ld';
             creating new 'RDF' graphs, and performing graph queries using 'SPARQL'.
             This package wraps the low level 'redland' R package which
             provides direct bindings to the 'redland' C library.  Additionally,
             the package supports the newer and more developer friendly
             'JSON-LD' format through the 'jsonld' package. The package
             interface takes inspiration from the Python 'rdflib' library.
Imports: redland,
    jsonld,
    utils
Suggests: magrittr,
    covr,
    httr,
    xml2,
    jqr,
    DT

```

https://github.com/cboettig/rdflib

data extraction, because this package parses scientific data file formats. (specifically, formats already parsed by existing rOpenSci packages).  This package also enables graph queries using the SPARQL language, somewhat analogous to the rOpenSci `jqr` package, but for JSON-LD and other linked data formats.

Anyone working with semantic data, including the wide array of scientific ontologies and knowledge-bases.  These include reproducibility-focused ontologies like PROV, and a large number of [biological ontologies](http://www.obofoundry.org/) ranging from genes to traits to environmental features.

As described above, this package overlaps significantly with the `redland` package, but should be easier to use.

    - [x] The package is deposited in a long-term repository with the DOI: https://doi.org/10.5281/zenodo.1098478
	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.

* Scott Chamberlain, @sckott
* Peter Slaughter, @gothub
* Bryce Mecum, @amoeba
* Anna Krystalli, @annakrystalli
