+++ DESCRIPTION
@@ -4,7 +4,7 @@ Version: 0.0.2
 Authors@R: person("Carl", "Boettiger", 
                   email = "cboettig@gmail.com", 
                   role = c("aut", "cre", "cph"),
-                  comment=c(ORCID = "http://orcid.org/0000-0002-1642-628X"))
+                  comment=c(ORCID = "0000-0002-1642-628X"))
 Description: The Resource Description Framework, or 'RDF' is a widely used
              data representation model that forms the cornerstone of the 
              Semantic Web. 'RDF' represents data as a graph rather than 
