### 2017-12-14 #169 @lmullen
### Editor checks:

---

#### Editor comments

Passes `devtools::check()` without issue.

```
R CMD check results
0 errors | 0 warnings | 0 notes
```
Here is the result of `goodpractice::gp()`.

```
It is good practice to

  ✖ write unit tests for all functions, and all
    package code in general. 97% of code lines are covered by
    test cases.

    R/rdf.R:274:NA
    R/rdf.R:275:NA

  ✖ not import packages as a whole, as this can cause
    name clashes between the imported packages. Instead, import
    only the specific functions you need.
```

97% coverage is excellent.

The advice not to import entire packages is up to the discretion of the reviewers.

Note one misspelling (`occured`) from `devtools::spell_check()`.

@karthik I am now approaching reviewers.

---

Reviewers:
Due date:
