### 2018-01-02 0700294
fix setting of a base uri in jsonld
uses localhost:// as base uri for now, configurable as option "rdflib_base_uri", though not documented. (since really users wanting a particular base_uri should probably be declaring their own base explicitly in the json-ld rather than setting it here).  closes #5
