+++ NAMESPACE
@@ -15,6 +15,8 @@ importClassesFrom(redland,Serializer)
 importClassesFrom(redland,Statement)
 importClassesFrom(redland,Storage)
 importClassesFrom(redland,World)
+importFrom(jsonld,jsonld_compact)
+importFrom(jsonld,jsonld_expand)
 importFrom(jsonld,jsonld_to_rdf)
 importFrom(utils,download.file)
 importMethodsFrom(redland,addStatement)
