### 2018-01-02 089989b
bump version, update news
- just import 'new' from methods, not whole package.
- do not skip_on_cran the tests requiring a url (so we get full credit for test coverage :rocket:)
