+++ NEWS.md
@@ -1,3 +1,10 @@
+# rdflib 0.0.3
+
+* add paper.md
+* add package level documentation
+* set base uri when serializing json-ld to rdf ([#5](https://github.com/cboettig/rdflib/issues/5))
+
+
 # rdflib 0.0.2
 
 * Added a `NEWS.md` file to track changes to the package.
