+++ NAMESPACE
@@ -7,7 +7,6 @@ export(rdf_add)
 export(rdf_parse)
 export(rdf_query)
 export(rdf_serialize)
-import(methods)
 importClassesFrom(redland,Model)
 importClassesFrom(redland,Parser)
 importClassesFrom(redland,Query)
@@ -18,6 +17,7 @@ importClassesFrom(redland,World)
 importFrom(jsonld,jsonld_compact)
 importFrom(jsonld,jsonld_expand)
 importFrom(jsonld,jsonld_to_rdf)
+importFrom(methods,new)
 importFrom(utils,download.file)
 importMethodsFrom(redland,addStatement)
 importMethodsFrom(redland,executeQuery)
