### 2018-01-02 4f0eae7
example of tabularizing triples, closes #8
