### 2018-01-03 #169 @lmullen
Two reviewers have agreed to review this package. Reviewers, thanks for being willing, and I'll ask you to have your reviews in within three weeks. Here is the [reviewer's guide](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md). Feel free to let me know if you have any questions.

Reviewer: Anna Krystalli, @annakrystalli
Reviewer: Bryce Mecum, @amoeba
Due date: 2018-01-25

@cboettig Could you please add the rOpenSci under review badge to the README for this package? Here is the snippet.
