+++ inst/examples/rdf_table.R
@@ -12,8 +12,8 @@ df
 
 
 ## Take a table of triples (columns: id, property, value)
-## and create a table with ids as rows and properties as columns (values as cell values)
-
+## and create a table with ids as rows and properties
+## as columns (values as cell values)
 spread_triples <- function(df){
   df %>%
     select(subject, predicate, object) %>%
@@ -21,7 +21,8 @@ spread_triples <- function(df){
     mutate(key = row_number()) %>%
     spread(predicate, object) %>%
     select(-key) %>%
-    fill(-subject) %>%  ## Fill in NA based on known properties for id, needs group_by id
+    ## Fill in NA based on known properties for id, needs group_by id
+    fill(-subject) %>%  
     ungroup()
 }
 
@@ -32,15 +33,25 @@ spread_triples()
 
 
 
-## a method based on parsing nquads, may not be the easiest serialization to tabularize.
+## a method based on parsing nquads, may not be 
+## the easiest serialization to tabularize.
 ## but works with just jsonld_to_rdf without any use of redland
 library(tidyverse)
 tabularize <- function(file){
   ## Express type string inside "" so we can parse
-  read_lines(file) %>% str_replace("\"\\^\\^(.*)>", "\\^\\^\\1\"") %>% write_lines(file)
-  df <- readr::read_delim(file, delim = " ", col_names = FALSE, col_types = "cccc")
+  read_lines(file) %>% 
+    str_replace("\"\\^\\^(.*)>", "\\^\\^\\1\"") %>%
+    write_lines(file)
+  df <- readr::read_delim(file, delim = " ", 
+                          col_names = FALSE, col_types = "cccc")
   names(df) <- c("id", "property", "value", "about")
   df %>%
-    mutate(property = gsub("<http://ecoinformatics.org/eml-2.1.1/(.*)>", "\\1", property)) %>%
-    separate(value, c("value", "type"), sep="\\^\\^<", fill="right")
+    mutate(property = 
+      gsub("<http://ecoinformatics.org/eml-2.1.1/(.*)>", 
+           "\\1", 
+           property)) %>%
+    separate(value, 
+             c("value", "type"), 
+             sep="\\^\\^<", 
+             fill="right")
 }
