### 2018-01-05 26dbcbc
use r-travis with addons
Notes: `librf0-dev` breaks r-travis when sudo is on.  (some conflict with Debian version of `r-cran-curl`; with travis `addons` we don't make sudo available, and thus r-travis just installs the R packages from source and bypasses this conflict.)
