+++ .travis.yml
@@ -1,6 +1,9 @@
-sudo: required
-services:
-  - docker
-script:
-  - docker run -ti -v $(pwd):/home/rstudio/repo -w /home/rstudio/repo rocker/verse:latest bash -c "apt-get update && apt-get -y install libjq-dev && R -e 'devtools::install(dep=TRUE); status <- devtools::check(); testthat::expect_length(status[[1]],0); testthat::expect_length(status[[2]],0)'"
-    
+# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
+language: R
+cache: packages
+addons:
+  apt:
+    packages:
+      - librdf0-dev
+after_success:
+  - Rscript -e 'covr::codecov()'
