### 2018-01-06 d4f7533
add v8 and libjq
