+++ .travis.yml
@@ -5,5 +5,7 @@ addons:
   apt:
     packages:
       - librdf0-dev
+      - libv8-dev
+      - libjq-dev
 after_success:
   - Rscript -e 'covr::codecov()'
