### 2018-01-06 c2e591b
drop sudo, can we add ppa?
