### 2018-01-06 2c150e2
Update DESCRIPTION
