### 2018-01-23 #169 @amoeba
>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

This looks like an excellent package for inclusion in the ropensci ecosystem. I have personal experience with the problem this package is trying to solve, namely working with RDF in R from the analysts perspective. The RDF package this package wraps, `redland`, is intended more for developers to write packages on top of than for end users to do things like run SPARQL queries or manipulate RDF graphs. I would use this package.

The package is laid out in a non-surprising manner, most functions are short and well-scoped, and, overall, the code is very readable. The accompanying test suite is reasonable and provides 100% coverage, and the single vignette is well-written and useful.

I did find the documentation could use some polish in some places (see comments below). I suspect a pass or two by the author would make some good improvements without much work.

I have left two checkboxes unchecked due to the following issues:

- **Community guidelines** are missing in README.md
- README.md missing citation section recommended by [packaging guide](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-readme)
- NEWS.md file doesn't quite conform to [packaging guide](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-news)

but I otherwise found everything else to be in order.

#### Higher level

- How are you dealing with differentiating between resource, literal, and blank nodes in `rdf_add`? As far as I can see, it looks I can't create a new triple with a resource or blank node as the subject:

  ```r
  > rdf_add(x, "test", "test", "test")
  > x
  <test> <test> "test" .
  ```

- I'd like to see more usage of the RDF capabilities of rdflib in the vignette. At current it's centered around JSON-LD

#### Lower level

These were written out as I went through the checklist:

- Documentation could benefit from a general pass for capitalization just to make things look nicer
- `\link{}`s in docs for `rdf_serialize` and `rdf_query` link to wrong version of the `parse` function
- Noticed funky encoding issues from vita.json.
  "Fernández" => "Fern\\u00E1ndez"
  Not sure if this is an rdflib problem or a redland one.
- I think any mentions of "RDF+XML" outside of MIME types should be "RDF/XML", not "RDF+XML"
- I thought arg `doc` in `rdf_serialize` could be `path` instead as it's a little more clear.
- I'm generally not a fan of arguments named `x`, though it feels warranted in some cases. Could a more descriptive name be used in these instances? For example, `rdf_query` has a first arg of `x` which could, instead, be `rdf`.
- Perhaps a little more error checking...
  If I send a malformed query, I get useful but a bit cryptic response from `rdf_query`:

  ```r
  > coi <- vita %>% rdf_query(sparql)
  librdf error  - syntax error, unexpected a
  rdf_query_results.c:100: (librdf_query_results_finished) assertion failed: object pointer of type librdf_query_results is NULL.
  rdf_query_results.c:100: (librdf_query_results_finished) assertion failed: object pointer of type librdf_query_results is NULL.
  ```

    This is caused by failing to cath an error to the call to `redland::executeQuery`.

    ```r
    queryResult <- redland::executeQuery(queryObj, x$model)
    ```

- `browseVignettes(package = "rdflib")` doesn't return any vignettes. I'm not that familiar with this functionality so I didn't debug it.
- Noticed the test "we can parse from a url" calls out to the web but the `skip_on_cran()` guard is commented out. Should it not be commented out?
- Found documentation for `rdf()` to be too terse and terse by comparison to other functions in the package
- `rdf_parse()` mentions the return is an 'rdf S3 object'. I might just call it an 'object' and omit the class system it was defined in
- Verbage in `rdf_serialize` 'Serialize RDF docs' maybe should be 'Serialize an RDF Documnet' (no plural)
- In `rdf_serialize`, I wonder if the return should not be the raw numeric return code from `redland` but either a logical or the document path itself (the latter would allow for piping)
- `rdf_query` states the return type is 'a list of all query results' when it actually appears to be a 'data.frame' from what I can tell
- What the common argument 'x' is documented as across functions varies. "rdf graph object" "rdf object". Please use common language for these
- RE: JOSS submission: I wonder if they won't find this package to simple/thin for acceptance. It does have clear research application though!
