### 2018-01-31 9f2518c
Working on revisions from rOpenSci onboarding.
