+++ NEWS.md
@@ -1,18 +1,28 @@
-# rdflib 0.0.3
+# rdflib 0.0.4
+
+
+# rdflib 0.0.3 (2018-01-02)
+
+## New Features
 
 * add paper.md
 * add package level documentation
+
+## Bug Fixes
+
 * set base uri when serializing json-ld to rdf ([#5](https://github.com/cboettig/rdflib/issues/5))
 
 
-# rdflib 0.0.2
+# rdflib 0.0.2 (2018-01-02)
+
+## New Features
 
 * Added a `NEWS.md` file to track changes to the package.
 * sparql query returns a data.frame format
 * added a vignette
 * added pkgdown website for vignette
 
-# rdflib 0.0.1
+# rdflib 0.0.1 (2017-12-09)
 
 * Initial prototype
 
