+++ R/rdf.R
@@ -1,7 +1,17 @@
 
-#' rdf constructor
+#' Initialize an `rdf` Object
 #'
 #' @return an rdf object
+#' @details an rdf Object is a list of class 'rdf', consisting of
+#' two pointers to external C objects managed by the redland library.
+#' These are the `World` object, basically a top-level pointer for
+#' all RDF models, and a `Model` object, essentially a storage structure
+#' for all RDF triples. `rdflib` defaults to an in-memory hash-based 
+#' storage structure at this time. The primary purpose of the `rdf`
+#' object is to abstract these low-level details away from the user.
+#' Typical use will be simply to initialize a container to which
+#' the user would manually add triples using \code{\link{rdf_add}}.
+#'
 #' @export
 #'
 #' @examples
@@ -35,7 +45,7 @@ print.rdf <- function(x, ...){
 }
 
 
-#' Parse RDF files
+#' Parse RDF Files
 #'
 #' @param doc path, URL, or literal string of the rdf document to parse
 #' @param format rdf serialization format of the doc,
@@ -79,12 +89,12 @@ rdf_parse <- function(doc,
     doc <- tmp
   }
   
-  x <- rdf()
+  rdf <- rdf()
   mimetype <- unname(rdf_mimetypes[format])
-  parser <- new("Parser", x$world, name = format, mimeType = mimetype)
-  redland::parseFileIntoModel(parser, x$world, doc, x$model)
+  parser <- new("Parser", rdf$world, name = format, mimeType = mimetype)
+  redland::parseFileIntoModel(parser, rdf$world, doc, rdf$model)
 
-  x
+  rdf
 }
 
 # Whenever we convert JSON-LD to RDF we should set a @base if not set.
@@ -116,7 +126,8 @@ add_base_uri <- function(doc, tmp = tempfile()){
 #' @param prefix string giving the prefix associated with the namespace
 #'
 #' @return rdf_serialize returns the output file path `doc` invisibly.
-#'   This makes it easier to use rdf_serialize in pipe chains with rdf_parse.
+#'   This makes it easier to use rdf_serialize in pipe chains with
+#'   \code{\link{rdf_parse}}.
 #' @importFrom methods new
 #' @importClassesFrom redland Serializer
 #' @importMethodsFrom redland setNameSpace serializeToFile
@@ -135,7 +146,7 @@ add_base_uri <- function(doc, tmp = tempfile()){
 #'           namespace = "http://purl.org/dc/elements/1.1/",
 #'           prefix = "dc")
 #'
-rdf_serialize <- function(x,
+rdf_serialize <- function(rdf,
                           doc,
                           format = c("rdfxml",
                                      "nquads",
@@ -159,18 +170,18 @@ rdf_serialize <- function(x,
   mimetype <- rdf_mimetypes[format]
 
   serializer <-
-    new("Serializer", x$world,
+    new("Serializer", rdf$world,
         name = format, mimeType = mimetype)
 
   if(!is.null(namespace)){
     redland::setNameSpace(serializer,
-                 x$world,
+                 rdf$world,
                  namespace = namespace,
                  prefix = prefix)
   }
  
   status <-
-    redland::serializeToFile(serializer, x$world, x$model, doc)
+    redland::serializeToFile(serializer, rdf$world, rdf$model, doc)
   
   if(jsonld_output){
     txt <- paste(readLines(doc), collapse = "\n")
@@ -180,14 +191,14 @@ rdf_serialize <- function(x,
     }
   }
   
-  invisible(x)
+  invisible(doc)
 }
 
 
 
-#' SPARQL query
+#' Perform a SPARQL Query
 #'
[…]
