### 2018-02-02 1e6b457
Work in progress:
- adds a c() method to concat rdf objects
- rdf_query() now coerces return type if it recognizes the data URI and can match it to an R type (a few XMLSchema types are recognized, otherwise defaults to character string).
- drop support for trig, which redland bindings fails to recognize
- fix support for turtle by changing mime=type for raptor
- add_rdf now coerces object to string type automatically
- experimental work to serialize arbitrary data.frames to RDF
