+++ man/c.rdf.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rdf.R
+\name{c.rdf}
+\alias{c.rdf}
+\title{Concatenate rdf Objects
+Note: this assumes absolute URIs for subject and predicate}
+\usage{
+\method{c}{rdf}(...)
+}
+\arguments{
+\item{...}{objects to be concatenated}
+}
+\description{
+Concatenate rdf Objects
+Note: this assumes absolute URIs for subject and predicate
+}
