### 2018-02-03 7c21c6b
support for BDB backend for disk-based storage, closes #6
Also extends tests, exports rdf_free, minor tweaks to documentation.  Attempt to install bdb on travis for testing as well.
