+++ tests/testthat/test-rdf.R
@@ -3,14 +3,54 @@ testthat::context("test-rdf.R")
 doc <- system.file("extdata/example.rdf", package="redland")
 out <- "testing.rdf"
 
+
+testthat::test_that("We can initialize and free rdf objects", {
+  rdf <- rdf()
+  
+  testthat::expect_is(rdf, "rdf")
+  testthat::expect_is(rdf$world, "World")
+  testthat::expect_is(rdf$model, "Model")
+  testthat::expect_is(rdf$storage, "Storage")
+  
+  rdf_free(rdf)
+})
+
+testthat::test_that("We warn if we cannot use disk-based storage", {
+  testthat::skip_if(has_bdb())
+  path <- tempdir()
+  testthat::expect_warning(rdf <- rdf(path), "BDB driver not found")
+  
+  ## Falls back on memory-based storage, still creates rdf
+  testthat::expect_is(rdf, "rdf")
+  rdf_free(rdf)
+  
+})
+
+
+
+testthat::test_that("We can use BDB storage", {
+  
+  testthat::skip_if_not(has_bdb())
+  path <- tempdir()
+  testthat::expect_silent(rdf <- rdf(path))
+  testthat::expect_is(rdf, "rdf")
+
+  rdf_free(rdf)
+  
+})
+
+
 testthat::test_that("we can parse (in rdfxml)
                     and serialize (in nquads) a simple rdf graph", {
   rdf <- rdf_parse(doc) 
   rdf_serialize(rdf, out, "nquads")
   roundtrip <- rdf_parse(out, "nquads")
   testthat::expect_is(roundtrip, "rdf")
+  
+  rdf_free(rdf)
 })
 
+## FIXME check format, check return types
 testthat::test_that("we can make sparql queries", {
   sparql <-
   'PREFIX dc: <http://purl.org/dc/elements/1.1/>
@@ -20,16 +60,35 @@ testthat::test_that("we can make sparql queries", {
   rdf <- rdf_parse(doc)
   match <- rdf_query(rdf, sparql)
   testthat::expect_length(match, 2)
+  
+  rdf_free(rdf)
+  
 })
 
 testthat::test_that("we can initialize add triples to rdf graph", {
-  x <- rdf()
-  x <- rdf_add(x, 
+  rdf <- rdf()
+  rdf <- rdf_add(rdf, 
            subject="http://www.dajobe.org/",
            predicate="http://purl.org/dc/elements/1.1/language",
            object="en")
-  testthat::expect_is(x, "rdf")
+  testthat::expect_is(rdf, "rdf")
+  rdf_free(rdf)
+  
 })
+
+
+testthat::test_that("we can concatenate rdfs", {
+  rdf1 <- rdf_parse(system.file("extdata/ex2.xml", package = "rdflib"))
+  rdf2 <- rdf_parse(doc)
+  rdf <- c(rdf1, rdf2)
+  testthat::expect_is(rdf, "rdf")
+  
+  rdf_free(rdf1)
+  rdf_free(rdf2)
+  rdf_free(rdf)
+})
+
+
 
 testthat::test_that("we can add, parse and serialize json-ld", {
   #x <- rdf_parse(doc)
@@ -41,7 +100,13 @@ testthat::test_that("we can add, parse and serialize json-ld", {
   rdf_serialize(x, out, "jsonld")
   rdf <- rdf_parse(out, format = "jsonld")
   testthat::expect_is(rdf, "rdf")
+  rdf_free(x)
+  rdf_free(rdf)
+  
 })
+
+
+
 
 testthat::test_that("print and format work", {
   rdf <- rdf_parse(doc)
@@ -49,15 +114,9 @@ testthat::test_that("print and format work", {
[…]
