+++ inst/examples/storage_types.R
@@ -1,17 +1,18 @@
+## Experimental testing of different storage backends in redland:
 
 library(redland)
 world <- new("World")
 
-## No error but null pointer returned
+## Works if BDB installed when we compile.  (not true on mac binaries from CRAN :/)
 bdb_storage <- new("Storage", world, "hashes", name = "db1", 
                       options = "new='yes',hash-type='bdb',dir='.'")
 model <- new("Model", world = world, storage = bdb_storage, options = "")
 
 ## error: sqlite not found
-sqlite_storage <- new("Storage", world, "sqlite", name = "sqlite1", options = "new='yes'")
+#sqlite_storage <- new("Storage", world, "sqlite", name = "sqlite1", options = "new='yes'")
 ## not found
-postgres_storage <- new("Storage", world, "postgresql", name = "postgres1", 
-    options = "new='yes',host='localhost',database='red',user='foo','password='bar'")
+#postgres_storage <- new("Storage", world, "postgresql", name = "postgres1", 
+#    options = "new='yes',host='localhost',database='red',user='foo','password='bar'")
 
 ## Works, in memory, serializes to an rdf/xml file called thing.rdf when freed.
 ## Not indexed, so will be slow. Suitable for small models.
@@ -26,11 +27,10 @@ storage <- memory_storage
 model <- new("Model", world = world, storage = storage, options = "")
 
 
+## Testing
 library(rdflib)
-
 rdf <- structure(list(world = world, model = model, storage = storage),
           class = "rdf")
-
 rdf_add(rdf, 
         subject="http://www.dajobe.org/",
         predicate="http://purl.org/dc/elements/1.1/language",
