### 2018-02-03 61d8833
a few notes to self
