### 2018-02-03 ab51029
royxgen docs use markdown :rocket: :sparkles: More examples of rdf_add() argument useage
