+++ NEWS.md
@@ -1,27 +1,39 @@
 # rdflib 0.1.0
 
+## New Features
+
 * add `c()` method to concatenate `rdf` objects
+* `rdf()` supports BDB backend for disk-based storage for large triplestores [#6](https://github.com/cboettig/rdflib/issues/6)
+
+## Minor Improvements
+
+* Numerous improvements to documentation, see 
+  [#9](https://github.com/cboettig/rdflib/issues/9) and 
+  [#10](https://github.com/cboettig/rdflib/issues/10) 
+
+## Bug Fixes 
+
 * `rdf_query` now coerces data into appropriate type 
    if it recognizes the data URI and can match that 
    to an R type (a few XMLSchema types are recognized,
    otherwise still defaults to character string)
-* All methods free memory from any temporary objects they initialize
+* Memory management: All methods free memory from any 
+  temporary objects they initialize, tests free memory.
   (e.g. parsers, serializers, query, statement)
-* rdf includes explicit pointer to storage object
-* rdf constructor supports BDB backend for disk-based triplestore [#6](https://github.com/cboettig/rdflib/issues/6)
-* tests free rdf objects
-* extend unit tests for some of new functionality
-* Add `rdf_free` to free rdf (ideally would be done by GC in redland...)
+* extend unit tests to cover new features
+* `turtle` parser/serializer fixed
+
+## Deprecated
+
+* `trig` support removed (was never working in redland)
+
 
 # rdflib 0.0.3 (2018-01-02)
 
-## New Features
+## Bug Fixes
 
 * add paper.md
 * add package level documentation
-
-## Bug Fixes
-
 * set base uri when serializing json-ld to rdf ([#5](https://github.com/cboettig/rdflib/issues/5))
 
 
