+++ DESCRIPTION
@@ -29,6 +29,7 @@ Imports: redland,
     methods,
     utils
 RoxygenNote: 6.0.1
+Roxygen: list(markdown = TRUE)
 Suggests: magrittr,
     covr,
     testthat,

[…]
