+++ inst/examples/as_rdf.R
@@ -49,29 +49,80 @@ xs_class <- function(x){
 
 
 as_rdf.list <- function(x){
-  
 }
 
 
 ## Testing: Digest some data.frames into RDF and extract back
-
+library(tidyverse)
 cars <- mtcars[1:4, 1:4] %>% rownames_to_column("Model")
-
-
 x1 <- as_rdf(iris)
 x2 <- as_rdf(cars)
 rdf <- c(x1,x2)
+
+
+#' @param ... names of the properties that should make up columns in the table
+#' @param columns Alternatively, supply a vector of property names. 
+#' @param prefix the URI string to prefix before the property names to give
+#'  fully-resolved properties.  
+#' @param na.rm logical, default TRUE. Will not return a row for any 
+#'  subject that does not have a object value matching the schema.
+#' @examples
+#' 
+#' sparql <- table_schema("Species",  "Sepal.Length", prefix = "iris")
+#' rdf_query(rdf, sparql)
+#' 
+#' ## use columns arg for an existing vector of names.
+#' columns <- names(iris)
+#' sparql <- table_schema(columns = columns, prefix = "iris")
+#' rdf_query(rdf, sparql)
+#' 
+#' ## use na.rm = FALSE to include NA if variable is not defined for some observations
+#' sparql <- table_schema("Species",  "Sepal.Length", "Sepal.Color",
+#'                        prefix = "iris", na.rm=FALSE)
+#' rdf_query(rdf, sparql)
+table_schema <- function(..., columns = NULL, prefix=NULL, na.rm = TRUE){
+  if(is.null(columns)){
+    columns <- c(...)
+  }
+  
+  if(!is.null(prefix)){
+    attributes <- paste0("<",prefix, ":", columns, ">")
+  } else {
+    attributes <- paste0("<", columns, ">")
+  }
+  ## Replace forbidden characters with "_", 
+  ## See spec: https://www.w3.org/TR/rdf-sparql-query/#rVARNAME
+  vars <- gsub("[^a-zA-Z1-9_]","_", basename(columns))
+  select <- paste0("?", vars)
+  
+  where <- paste("?s", attributes, select)
+  if(!na.rm)
+    where <- paste("OPTIONAL {", where, "}")
+  else
+    where <- paste(where, ".")
+  where <- paste(where, collapse = "\n")
+  query <- paste("SELECT", paste0(select, collapse = " "), "\nWHERE {\n", where, "\n}")
+  query
+}
+
+  
+sparql <- table_schema("Sepal.Length", "Sepal.Width", 
+                       "Petal.Length", "Petal.Width",  
+                       "Species", prefix="iris")
+sparql <- table_schema(columns=names(iris), prefix="iris")
+rdf_query(rdf, sparql)
+
 
 ## Look, original data back, no filter / spread required!
 ## Could write helper function to construct this pattern of sparql?
 ## e.g. return a data.frame with all observations of a list of attributes (properties)
 sparql <-
-  'SELECT ?Sepal_Length ?Sepal_Width ?Petal_Length ?Petal_Width ?Species
+  'SELECT ?Sepal_Length ?Sepal_Width ?Petal_Length ?Petal_Color ?Species
    WHERE {
-      ?s <iris:Sepal.Width>  ?Sepal_Width . 
-      ?s <iris:Sepal.Length>  ?Sepal_Length . 
-      ?s <iris:Petal.Width>  ?Petal_Width . 
-      ?s <iris:Petal.Length>  ?Petal_Length . 
-      ?s <iris:Species>  ?Species . 
[…]
