+++ NAMESPACE
@@ -21,6 +21,7 @@ importFrom(jsonld,jsonld_expand)
 importFrom(jsonld,jsonld_to_rdf)
 importFrom(methods,as)
 importFrom(methods,new)
+importFrom(stringi,stri_unescape_unicode)
 importFrom(utils,capture.output)
 importFrom(utils,download.file)
 importMethodsFrom(redland,addStatement)

[…]
