### 2018-02-05 73a0e94
Move defs for exported funs into separate R files
Also: rdf_add will automatically datatype
