+++ R/utilities.R
@@ -1,3 +1,10 @@
+#' @importFrom methods as
+type_by_datauri <- function(x){
+  types <- get_types(x)
+  r_types <- vapply(get_types(x), r_class, character(length(1)))
+  df <- data.frame(value = get_values(x), class = r_types)
+  apply(df, 1, function(x) as(utf8me(x[1]), x[2]))
+}
 
 ## Utilities to coerce return type, if recognized
 r_class <- function(x){
@@ -12,33 +19,19 @@ r_class <- function(x){
          "character"
   )
 }
+
+## Helper utilitiesfor parsing data URIs
 get_values <- function(x) gsub('\"(([^\\^])+)\"\\^*(.*)',  "\\1", x)
 get_types <- function(x) out <- gsub('\"(([^\\^])+)\"\\^*(.*)',  "\\3", x)
 
-
-#' @importFrom methods as
-type_by_datauri <- function(x){
-  types <- get_types(x)
-  r_types <- vapply(get_types(x), r_class, character(length(1)))
-  df <- data.frame(value = get_values(x), class = r_types)
-  apply(df, 1, function(x) as(utf8me(x[1]), x[2]))
-}
-
 #' @importFrom stringi stri_unescape_unicode
+# https://stackoverflow.com/questions/48602294
 utf8me <- function(x){ 
   removed_quotes <- gsub('\"', '', x)
   stringi::stri_unescape_unicode(removed_quotes)
 }
 
-rectangularize_query_results <- function(out){
-  vars <- unique(names(out))
-  X <- lapply(vars, function(v){ 
-     contents <- as.character(out[names(out) == v ])
-     type_by_datauri(contents)
-  })
-  names(X) <- vars
-  as.data.frame(X, stringsAsFactors=FALSE)
-}
+
 
 
 
@@ -79,17 +72,4 @@ rdf_mimetypes <- c("nquads" = "text/x-nquads",
 # - rss (application/rss+xml or text/rss)
 
 
-# rdf functions like working with local files
-# this helper function allows us to also use URLs or strings
-#' @importFrom utils download.file
-text_or_url_to_doc <- function(x, tmp = tempfile()){
-  if(file.exists(x)){
-    return(x) 
-  } else if(grepl("^https?://", x)) {
-    utils::download.file(x, tmp, quiet = TRUE)
-    return(tmp)
-  } else {
-    writeLines(x, tmp)
-    return(tmp)
-  }
-}
+
