+++ man/rdf_query.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rdf.R
+% Please edit documentation in R/rdf_query.R
 \name{rdf_query}
 \alias{rdf_query}
 \title{Perform a SPARQL Query}
