+++ man/rdf_serialize.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rdf.R
+% Please edit documentation in R/rdf_serialize.R
 \name{rdf_serialize}
 \alias{rdf_serialize}
 \title{Serialize an RDF Document}
