+++ man/rdf_parse.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rdf.R
+% Please edit documentation in R/rdf_parse.R
 \name{rdf_parse}
 \alias{rdf_parse}
 \title{Parse RDF Files}

[…]
