### 2018-02-06 aef4728
Reorganize & bugfix * separate files for each main function, separate files for tests * fix issue with c.rdf / format.rdf on string escapes * fix Date/time handling in rdf_add / rdf_query * extend documentation * extend tests for type handling
