+++ inst/examples/debug.R
@@ -51,3 +51,35 @@ predicate="http://schema.org/name",
 object="Maëlle Salmon")
 
 r
+
+
+x <- "\"2018-02-05 11:54:10\"^^<http://www.w3.org/2001/XMLSchema#dateTime>"
+type_by_datauri(x)
+
+reprex::reprex({
+  tbl <- tibble:::as_data_frame(list(col1  = list(Sys.Date(), "stuff")))
+  df <- as.data.frame(tbl)
+  df
+  
+  df[1,1]
+})
+
+
+get_values(x)
+get_types(x)
+
+
+
+
+
+a <- '\"some text\"'
+b <- '"some text"'
+identical(a,b) #TRUE
+
+txt <- "\"Ma\\u00EBlle\""
+stringi::stri_unescape_unicode(txt)
+
+# match only interior quotes:
+a <- '"some text with \"interior quotes\" and more text"'
+gsub("(^.+(\").+$)", "\\2", a)
+stringr::str_replace(a, )

[…]
