+++ inst/examples/as_rdf.R
@@ -1,27 +1,26 @@
 as_rdf <- function(df, base_uri = NULL) UseMethod("as_rdf")
 
-as_rdf.data.frame <- function(df, base_uri = NULL){
 
-  subjectType <- "blank"
-  if(is.null(base_uri)){
-    ## Use literal name of the object passed to the function as df
-    base_uri <- paste0(gsub("[^a-zA-Z]","",deparse(substitute(df))), ":")
-    subjectType <- "uri"
-  }
+## tidy data to rdf
+as_rdf.data.frame <- function(df, base_uri = NULL){
   
-  ## better to avoid heavy use of tidyverse?  All but `tidyr::gather` should
-  ## be avoidable
   x <- tibble::rowid_to_column(df, "subject")
   suppressWarnings(
     x <- tidyr::gather(x, key = predicate, value = object, -subject)
   )
+  
+  ## gather looses col-classes
+  col_classes <- data.frame(datatype = vapply(df, rdflib:::xs_class, character(1)))
+  col_classes <- tibble::rownames_to_column(col_classes, "predicate")
+  x <- dplyr::inner_join(x, col_classes, "predicate")
   
   rdf <- rdf()
   for(i in seq_along(x$subject)){
     rdf <- rdf_add(rdf, 
                    subject = paste0(base_uri, as.character(x$subject[[i]])),
                    predicate = paste0(base_uri, x$predicate[[i]]),
-                   object = x$object[[i]])
+                   object = x$object[[i]],
+                   datatype_uri = x$datatype[[i]])
   }
   rdf
 }
@@ -33,17 +32,21 @@ as_rdf.list <- function(x){
 
 
 ## Testing: Digest some data.frames into RDF and extract back
-library(tidyverse)
-cars <- mtcars[1:4, 1:4] %>% rownames_to_column("Model")
-x1 <- as_rdf(iris)
-x2 <- as_rdf(cars)
-rdf <- c(x1,x2)
+# library(tidyverse)
+# cars <- mtcars %>% rownames_to_column("Model")
+# x1 <- as_rdf(iris, "iris:")
+# x2 <- as_rdf(cars, "mtcars:")
+# rdf <- c(x1,x2)
+
+rdf_serialize(x1, "iris.json", "jsonld")
+jsonld::jsonld_compact("iris.json", context = '{"@base": "iris:"}')
 
 
-
-
-
-
+#' Tidy Schema
+#' 
+#' Construct a SPARQL query in which properties (predicates) become column headings
+#' and values (objects) become entries of individual cells.  
+#' 
 #' @param ... names of the properties that should make up columns in the table
 #' @param columns Alternatively, supply a vector of property names. 
 #' @param prefix the URI string to prefix before the property names to give
@@ -64,7 +67,7 @@ rdf <- c(x1,x2)
 #' sparql <- table_schema("Species",  "Sepal.Length", "Sepal.Color",
 #'                        prefix = "iris", na.rm=FALSE)
 #' rdf_query(rdf, sparql)
-table_schema <- function(..., columns = NULL, prefix=NULL, na.rm = TRUE){
+tidy_schema <- function(..., columns = NULL, prefix=NULL, na.rm = TRUE){
   if(is.null(columns)){
     columns <- c(...)
   }
@@ -90,24 +93,11 @@ table_schema <- function(..., columns = NULL, prefix=NULL, na.rm = TRUE){
   query
 }
 
-  
-sparql <- table_schema("Sepal.Length", "Sepal.Width", 
-                       "Petal.Length", "Petal.Width",  
-                       "Species", prefix="iris")
[…]
