+++ vignettes/rdf_intro.Rmd
@@ -18,66 +18,13 @@ In the world of data science, RDF is a bit of an ugly duckling.  Like XML and Ja
 
 [^1]: "The semantic web is the future of the internet and always will be." -Peter Norvig, Director of Research at Google
 
-But don't let these appearances deceive you. RDF really is cool. If you've ever gotten carried away using `tidyr::gather` to make everything into one long table, you may have noticed you can just about always get things down to about three columns, as we see in the obligatory `iris` data example in the `tidyr::gather` docs:
+But don't let these appearances deceive you. RDF really is cool. If you've ever gotten carried away using `tidyr::gather` to make everything into one long table, you may have noticed you can just about always get things down to about three columns, as we see with an obligatory `mtcars` data example for `tidyr::gather`:
 
 ```{r message = FALSE, warning=FALSE}
 library(rdflib)
 library(tidyverse)
 ```
 
-
-```{r}
-# get first observation for each Species in iris data -- base R
-mini_iris <- iris[c(1, 51, 101), ]
-# gather Sepal.Length, Sepal.Width, Petal.Length, Petal.Width
-iris_triples <- 
-  gather(mini_iris, key = attribute, value = measurement, -Species)
-iris_triples
-```
-
-If you like long tables like this, RDF is for you. This layout isn't "Tidy Data," where rows are observations and columns are variables, but it is damn useful sometimes.  This format is very liquid, easy to reshape into other structures -- so much so that `tidyr::gather` was originally known as `melt` in the `reshape2` package. 
-
-Looking at this table closely, we see that each row is reduced to the most elementary statement you can make from the data. A row no longer tells you the measurements (observations) *all* attributes (variables) of a given species (key), instead, you get just one fact per row, `setosa` has `Sepal.Length` measurement of `5.1`.  In RDF-world, we think of these three columns as something very special, which we call **triples**.  RDF is all about these triples. 
-
-The first column, `Species`, is the `key` from data.frame, i.e. the **subject** being described. The next column is the variable or attribute being measured, (that is, column names, other than the key column(s), from the tidy data), called the property or **predicate** in RDF-speak (slash grammar-school jargon). The third column is the actual value measured, more **object** of the predicate. Call it key-attribute-value or subject-predicate-object, these are our triples.  
-
-
-
-
-## Subject URIs
-
-In the examples we just considered, we have taken the liberty of treating a specific column (Species; Car model name) as the data key.  This was inutitive but actually a bit sloppy: for instance, we did not verify that these names are unique. We would be better off being more careful about this by treating the row IDs themselves as the `key` column for the data.frame (i.e. as the `subject` in our triple): 
-
-
-```{r}
-mini_iris %>%
-  rowid_to_column("subject") %>%
-  gather(key = predicate, value = object, -subject) %>% 
-  tail()
-```
-
-As you can see, this allows us to `gather` *all* columns, `Species` column included.  Our `subject` column now consists only of simple numeric `id`'s, while we have gained an additional triple for every row in the original data which states `Species` of each `id` (e.g. `1` is `Species` `setosa`).  
-## Predicate URIs
-
-Another more pernicious problem we may run into is that our different tables may use the same attribute (i.e. originally, a column name of a variable) for different things. 
-
-
-
-Okay, at this point you can probably start to spot a few possible issues.
-
-## RDF types
-
[…]
