### 2018-02-06 ed9df19
table of equivalent terminology
