+++ vignettes/rdf_intro.Rmd
@@ -3,8 +3,7 @@ title: "An tidyverse lover's introduction to RDF"
 author: "Carl Boettiger"
 date: "`r Sys.Date()`"
 output: 
-  rmarkdown::html_vignette:
-    df_print: paged
+  rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{rdflib Introduction}
   %\VignetteEngine{knitr::rmarkdown}
@@ -31,9 +30,11 @@ car_triples <-
 mtcars %>% 
   rownames_to_column("Model") %>% 
   gather(attribute,measurement, -Model)
-car_triples
 ```
 
+```{r echo=FALSE}
+DT::datatable(car_triples)
+```
 
 If you like long tables like this, RDF is for you. This layout isn't "Tidy Data," where rows are observations and columns are variables, but it is damn useful sometimes.  This format is very liquid, easy to reshape into other structures -- so much so that `tidyr::gather` was originally known as `melt` in the `reshape2` package. It's also a good way to get started thinking about RDF.
 
@@ -44,15 +45,17 @@ Looking at this table closely, we see that *each row is reduced to the most elem
 The first column came from the row names in this case, the `Model` of car.  This acts serves as a `key` to index the data.frame, i.e. the **subject** being described. The next column is the variable (also called attribute or property) being measured, (that is, column names, other than the key column(s), from the tidy data), called the property or **predicate** in RDF-speak (slash grammar-school jargon). The third column is the actual value measured, more **object** of the predicate. Call it key-property-value or subject-predicate-object, these are our triples.  We can represent just about any data in fully elementary manner.  
 
 
+&nbsp;          | &nbsp;  | &nbsp;      | &nbsp;
+----------------|---------|-------------|-------------
+**RDF**         | subject | predicate   | object         
+**JSON**        | object  | property    | value
+**spreadsheet** | row id  | column name | cell
+**data.frame**  | key     | variable    | measurement
+**data.frame**  | key     | attribute   | value  
 
-subject | predicate | object
-object  | property  | value
-row id  | column name | cell
-key     | variable  | measurement
-key     | attribute | value  
+Table: **Table 1**: The many names for triples.  
 
-
-Table: the many names for triples.  The first naming convention is the terminology typically associated with RDF.  The second set are terms typically associated with JSON data, while the remaining are all examples in tabular or relational data structures.  
+Table 1 summarizes the many different names associated with triples.  The first naming convention is the terminology typically associated with RDF.  The second set are terms typically associated with JSON data, while the remaining are all examples in tabular or relational data structures.  
 
 
 ## Subject URIs
@@ -64,30 +67,43 @@ car_triples <-
 mtcars %>% 
   rownames_to_column("Model") %>% 
   rowid_to_column("subject") %>% 
-
   gather(predicate, object, -subject)
-car_triples
 ```
+
+
+```{r echo=FALSE}
+DT::datatable(car_triples)
+```
+
 
[…]
