### 2018-02-12 5922ca1
a few features/improvements:
* Use `options` to control rdf storage
* rdf_parse can parse into an existing model
* streamline c() method.

closes #7
