+++ R/utilities.R
@@ -43,7 +43,19 @@ setAs("character", "Date", function(from) as.Date(as.character(from)))
 setAs("character", "POSIXct", function(from) as.POSIXct(as.character(from)))
 
 
-has_bdb <- function(){
+#' Check for BDB support
+#' 
+#' Detect whether Berkeley Database for disk-based storage of RDF graphs
+#' is available.  Disk-based storage requires redland package
+#' to be installed from source with support for the Berkeley DB 
+#' (libdb-dev on Ubuntu, berkeley-db on homebrew), otherwise `rdf()` will
+#' fall back to in-memory storage with a warning.
+#' 
+#' @return TRUE if BDB support is detected, false otherwise
+#' @export
+#' @examples 
+#' rdf_has_bdb()
+rdf_has_bdb <- function(){
   ## Unfortunately convoluted way to check if we have Berkeley DB Support
   world <- new("World")
   path <-tempdir()
