+++ man/rdf_has_bdb.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utilities.R
+\name{rdf_has_bdb}
+\alias{rdf_has_bdb}
+\title{Check for BDB support}
+\usage{
+rdf_has_bdb()
+}
+\value{
+TRUE if BDB support is detected, false otherwise
+}
+\description{
+Detect whether Berkeley Database for disk-based storage of RDF graphs
+is available.  Disk-based storage requires redland package
+to be installed from source with support for the Berkeley DB
+(libdb-dev on Ubuntu, berkeley-db on homebrew), otherwise \code{rdf()} will
+fall back to in-memory storage with a warning.
+}
+\examples{
+rdf_has_bdb()
+}
