+++ NAMESPACE
@@ -6,6 +6,7 @@ S3method(print,rdf)
 export(rdf)
 export(rdf_add)
 export(rdf_free)
+export(rdf_has_bdb)
 export(rdf_parse)
 export(rdf_query)
 export(rdf_serialize)

[…]
