### 2018-02-12 29a49cd
use separate files for sources of all exported functions
