+++ man/c.rdf.Rd
@@ -1,9 +1,8 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rdf.R
+% Please edit documentation in R/rdf_methods.R
 \name{c.rdf}
 \alias{c.rdf}
-\title{Concatenate rdf Objects
-Note: this assumes absolute URIs for subject and predicate}
+\title{Concatenate rdf Objects.}
 \usage{
 \method{c}{rdf}(...)
 }
@@ -11,6 +10,6 @@ Note: this assumes absolute URIs for subject and predicate}
 \item{...}{objects to be concatenated}
 }
 \description{
-Concatenate rdf Objects
-Note: this assumes absolute URIs for subject and predicate
+All subsequent rdf objects will be appended to the first rdf object
+Note: this does not free memory from any of the individual rdf objects
 }
