+++ man/rdf_has_bdb.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utilities.R
+% Please edit documentation in R/rdf_has_bdb.R
 \name{rdf_has_bdb}
 \alias{rdf_has_bdb}
 \title{Check for BDB support}
