+++ man/rdf_free.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rdf.R
+% Please edit documentation in R/rdf_free.R
 \name{rdf_free}
 \alias{rdf_free}
 \title{Free Memory Associated with RDF object}
