### 2018-02-13 7513bc2
rdf_free removes object from parent.frame
