+++ man/rdf_serialize.Rd
@@ -21,7 +21,7 @@ or "jsonld"}
 
 \item{prefix}{string giving the prefix associated with the namespace}
 
-\item{...}{additional parameters (not implemented)}
+\item{...}{additional arguments to \code{\link{redland::serializeToFile}}}
 }
 \value{
 rdf_serialize returns the output file path \code{doc} invisibly.
