### 2018-02-13 49365f7
add as_rdf.list() method
