### 2018-02-13 cdfd6d1
update NEWS
