+++ NEWS.md
@@ -2,35 +2,45 @@
 
 ## New Features
 
-* add `c()` method to concatenate `rdf` objects
-* `rdf()` supports BDB backend for disk-based storage for large triplestores [#6](https://github.com/cboettig/rdflib/issues/6)
+* adds `c()` method to concatenate `rdf` objects
+* `rdf()` supports BDB backend for disk-based storage for large
+   triplestores [#6](https://github.com/cboettig/rdflib/issues/6)
+* `rdf_parse()` gains an argument `rdf` to append triples to existing graph
 
 ## Minor Improvements
 
-* Numerous improvements to documentation, see 
-  [#9](https://github.com/cboettig/rdflib/issues/9) and 
-  [#10](https://github.com/cboettig/rdflib/issues/10) 
+
 * `rdf_add` argument for `object` can now take all atomic types
    (numeric, integer, string, Date, POSIX, logical) and 
    will automatically declare the appropriate `datatype_uri`
    if the user has not manually specified this. 
+* Numerous improvements to documentation from rOpenSci onboarding feedback, see 
+  [#9](https://github.com/cboettig/rdflib/issues/9) and 
+  [#10](https://github.com/cboettig/rdflib/issues/10) 
+* both functions and unit tests are broken out into separate files in
+  their respective directories.
+* additional example RDF data added in `extdata`
+* `rdf_serialize` passes `...` arguments to serializeToFile (e.g. to set a `baseUri`) 
 
 ## Bug Fixes 
 
+* `rdf_free()` will also remove the object from the parent frame, 
+  reducing the potential for crashing R by referring to a freed pointer.
 * fix encoding with UTF-8 characters (coming from nquads & ntriples)
-* `rdf_query` now coerces data into appropriate type 
+* `rdf_query()` now coerces data into appropriate type 
    if it recognizes the data URI and can match that 
    to an R type (a few XMLSchema types are recognized,
    otherwise still defaults to character string)
 * Memory management: All methods free memory from any 
   temporary objects they initialize, tests free memory.
   (e.g. parsers, serializers, query, statement)
-* extend unit tests to cover new features
+* extend unit tests to cover new features, check UTF-8
 * `turtle` parser/serializer fixed
 
 ## Deprecated
 
-* `trig` support removed (was never working in redland)
+* `trig` support removed (not working in redland without optional
+   libraries and alternative compile configuration)
 
 
 # rdflib 0.0.3 (2018-01-02)
