### 2018-02-13 d751f82
WIP: more on serializations, graphs
