+++ vignettes/rdf_intro.Rmd
@@ -11,7 +11,9 @@ vignette: >
   
 ---
 
-
+```{r include=FALSE}
+options(rdf_print_format = "nquads")
+```
 
 In the world of data science, RDF is a bit of an ugly duckling.  Like XML and Java, only without the massive-adoption-that-refuses-to-die part.  In fact RDF is most frequently expressed in XML, and RDF tools are written in Java, which help give RDF has the aesthetics of *steampunk*, of some technology for some futuristic Semantic Web[^1] in a toolset that feels about as lightweight and modern as iron dreadnaught.
 
@@ -22,6 +24,7 @@ But don't let these appearances deceive you. RDF really is cool. If you've ever 
 ```{r message = FALSE, warning=FALSE}
 library(rdflib)
 library(tidyverse)
+library(jsonld)
 ```
 
 
@@ -199,19 +202,149 @@ Note that we get a blank node, `_:`  with a randomly generated string.
 
 ## Triple notation: `nquads` `rdfxml`, `turtle`, and `nquads`
 
-So far we have relied primarily on a three-column tabular format to represent our triples.  
+So far we have relied primarily on a three-column tabular format to represent our triples.  We have also seen the default `print` format used for the `rdf` method, known as [N-Quads](https://www.w3.org/TR/n-quads/) above, which displays a bare, space-separated triple, possibly with a datatype URI attached to the object.  The line ends with a dot, which indicates this is part of the same local triplestore (aka RDF graph or RDF Model).  Technically this could be another URI indicating a unique global address for the triplestore in question.   
+
+We can serialize any `rdf` object out to a file in this format with the `rdf_serialize()` function, e.g.
+
+```{r}
+rdf_serialize(rdf, "rdf.nq", format = "nquads")
+```
+
+ Just as each of these formats can be serialized with `rdf_serialize()`, each can be read by `rdflib` using the function `rdf_parse()`:
+
+
+```{r}
+doc <- system.file("extdata/example.rdf", package="redland")
+rdf <- rdf_parse(doc, format = "rdfxml") 
+rdf
+```
+
+
+N-Quads are convenient in that each triple is displayed on a unique line, and the format supports the blank node and Datatype URIs in the manner we have just discussed. Other formats are not so concise.  Rather than print to file, we can simply change the default print format used by `rdflib` to explore the textual layout of the other serializations.  Here is one of the most common classical serializations, `RDF/XML` which expresses triples in an XML-based schema:
 
 ```{r}
 options(rdf_print_format = "rdfxml")
 rdf
 ```
 
+Just looking at this is probably enough to explain why so many alternative serializations were created.  Another popular format, [`turtle`](https://www.w3.org/TR/turtle/), looks more like `nquads`: 
 
 
-- JSON-LD
-- Reasoning
+```{r}
+options(rdf_print_format = "turtle")
+rdf
+```
 
-JSON-LD offers a fresh look at how we might serialize RDF in a way that is both familiar and easy to read.  The simplicity is really quite brilliant.  Just about all the ugly parts -- URI strings, type strings, even the prefixes -- are pushed up into the "context", a kind of header for holding these details, while stuff most applications will really care about is unencumbered, standard JSON.  This approach makes brilliant use of the "Object Notation" in the JSON to handle the ever-troublesome notion of a subject
+Here, blank nodes are denoted by `[]`.  `turtle` uses indentation to indicate that all three predicates (`creator`, `description`, `title`) are properties of the same subject.  
+
+
+
+### JSON-LD
+
[…]
