+++ DESCRIPTION
@@ -20,7 +20,6 @@ Description: The Resource Description Framework, or 'RDF' is a widely used
              'JSON-LD' format through the 'jsonld' package. The package
              interface takes inspiration from the Python 'rdflib' library.
 License: MIT + file LICENSE
-Date: 2018-02-12
 Encoding: UTF-8
 LazyData: true
 URL: https://github.com/cboettig/rdflib
