### 2018-02-14 #169 @cboettig
@lmullen @annakrystalli @amoeba Thanks for your reviews!

 I've just about finished addressing the issues raised at this point, which I've summarized in:

- [Bryce's review, issue #9](https://github.com/cboettig/rdflib/issues/9)
- [Anna's review, issue #10](https://github.com/cboettig/rdflib/issues/10)

A summary of the changes can be found in [NEWS.md](https://github.com/cboettig/rdflib/blob/master/NEWS.md), which ended up being reasonably involved because the reviews got me thinking about a bunch of stuff, which was awesome.

However, most substantive is perhaps the development of a new vignette, which I've liberally titled [A tidyverse lover’s intro to RDF](https://cboettig.github.io/rdflib/articles/rdf_intro.html).  This tries to address the big-picture issues Anna in particular highlights regarding documenting and motivating the broader context of RDF.  This is still a bit more of a draft than a polished document, but given that my two weeks are up I think it might be a good time to get feedback on this (and the other changes) from the reviewers.  In particular, I would love to hear what the reviewers think of this as a broader introduction.

If the reviewers are interested and think it would be worthwhile, I believe it might be nice to overhaul this new vignette into a more general purpose intro to RDF for R users (both the relevant packages and concepts) that might be suitable for a submission to something like the R Journal.  I'd love entice Anna and Bryce to be co-authors if they are interested...
