### 2018-02-15 #169 @annakrystalli
Hi all 😃

I am really happy with the changes made and the direction of the vignette! triplestores are indeed the ultimate tidy data! A great way to sell it. It's already a great resource and am also happy to contribute to both the vignette and a paper on it.  I'll feedback to some of the discussions raised in @cboettig response in`rdflib` [issue](https://github.com/cboettig/rdflib/issues/10).

So ✅ and big 👍 from me also.
