### 2018-02-15 14b7ebe
Use built-in base uri mechanism from jsonld
Much much faster (and simpler) than manual expansion with `@base`.  Follow-up on #5
