+++ vignettes/rdf_intro.Rmd
@@ -353,8 +353,8 @@ source(system.file("examples/as_rdf.R", package="rdflib"))
 ## Testing: Digest some data.frames into RDF and extract back
  library(tidyverse)
  cars <- mtcars %>% rownames_to_column("Model")
- x1 <- as_rdf(iris, "iris:")
- x2 <- as_rdf(cars, "mtcars:")
+ x1 <- as_rdf(iris, NULL, "iris:")
+ x2 <- as_rdf(cars, NULL, "mtcars:")
  rdf <- c(x1,x2)
 ```
 
