### 2018-02-15 6ed90d6
update docs, experimental as_rdf method
