+++ NAMESPACE
@@ -17,6 +17,7 @@ importClassesFrom(redland,Serializer)
 importClassesFrom(redland,Statement)
 importClassesFrom(redland,Storage)
 importClassesFrom(redland,World)
+importFrom(jsonld,jsonld_compact)
 importFrom(jsonld,jsonld_to_rdf)
 importFrom(methods,as)
 importFrom(methods,new)
