### 2018-02-16 a97a0dc
Performance stress-test with millions of triples
