+++ inst/examples/profile_performance.R
@@ -1,11 +1,9 @@
-
 library(nycflights13)
 library(tidyverse)
 library(rdflib)
 source(system.file("examples/as_rdf.R", package="rdflib"))
 
-
-## Must use BDB to successfully import flights 
+## Must use BDB to successfully import all of flights 
 options(rdflib_storage = "BDB")
 
 ## Tidyverse Style
@@ -14,30 +12,25 @@ df <- flights %>%
   left_join(planes, by="tailnum") %>% 
   select(carrier, manufacturer, model)
 
-## JSON-LD approach -- < 1 sec, 
-## ~ all rdf_parse
-## ~ all of which is inside jsonld_to_rdf()
-system.time(
-  x3 <- as_rdf.list(planes)
-)
 
+## generic list-based conversion via JSON-LD 
+rdf_planes_from_list <- as_rdf.list(planes)
 
-## rdf_add approach 33 sec
-## ~ all of which is rdf_add
-## ~ all of which is calls to S4 new() constructor
-#3 ~ none of which is calls to the low level librdf_* C calls
+as_uri <- function(x, base_uri = "x:") paste0("<", base_uri, x, ">")
 
+if(!rdf_has_bdb()){
+flights <- flights %>% 
+  filter(distance > 2500) # try smaller dataset
+}
+
+flights <- flights %>% 
+  mutate(tailnum = as_uri(tailnum),
+         carrier = as_uri(carrier)) -> flights
 
 ## New method, poor-mans-quads, < 0.5s, mostly gather
-system.time(
-x3 <- as_rdf(planes,  "tailnum", "x:")
-)
-
 x1 <- as_rdf(airlines, "carrier", "x:")
-
-
 x2 <- as_rdf(airports, "faa", "x:") ## a few funny chars, UTF8 issues?
-
+x3 <- as_rdf(planes,  "tailnum", "x:")
 
 
 sparql <-
@@ -47,39 +40,72 @@ WHERE {
  ?tailnum <x:model>  ?model .
 }'
 
-profvis::profvis(
-  out <- rdf_query(x3, sparql)
-)
+out <- rdf_query(x3, sparql)
 
 
-
-# 165 seconds
+# 165 seconds if this is the full table
 system.time(
     x4 <- as_rdf(flights, NULL, "x:")
 )
 
-## Nope, appears to crash R...
+## nope, the jsonld method appears to crash R...
 #system.time(
 #  x4 <- as_rdf.list(na.omit(flights))
 #)
 
+## we can join all of these:  
 rdf <- c(x4,x1,x2,x3)
 
-## FIXME: foreign keys must be URIs
+#}
 
 
 
+## separate queries:
+sparql <-
+'SELECT  ?tailnum ?dep_delay ?carrier
+WHERE { 
+  ?flight <x:tailnum>  ?tailnum .
+  ?flight <x:dep_delay>  ?dep_delay .
+  ?flight <x:carrier>  ?carrier 
+}'
+f1 <- rdf_query(rdf, sparql)
 
 sparql <-
-'SELECT  ?carrier ?tailnum ?dep_delay
+  'SELECT  ?tailnum ?model ?manufacturer
 WHERE {
- ?s <x:tailnum>  ?tailnum .
- ?s <x:dep_delay>  ?dep_delay .
- ?s <x:carrier>  ?carrier 
+?tailnum <x:manufacturer> ?manufacturer .
+?tailnum <x:model> ?model
+}'
+f2 <- rdf_query(rdf, sparql)
 
+tmp <- inner_join(f1,f2)
+
+## Can we get this to work in one single sparql Query?
[…]
