+++ inst/examples/profile_performance.Rmd
@@ -1,0 +1,211 @@
+
+```{r}
+library(nycflights13)
+library(tidyverse)
+library(rdflib)
+```
+
+
+
+```{r}
+source(system.file("examples/as_rdf.R", package="rdflib"))
+```
+
+
+We need to set `rdflib` option to use disk-based rather than in-memory storage,
+or it appears that `redland` throws an error (even when the machine has sufficient memory!?)
+when importing the 336,776 rows of the `flights` table.  
+
+Note: if BDB is not available (e.g. `berkeley-db` libraries were not found when `redland` was built from source),
+then this will fallback on in-memory storage and this vignette will use an abridged version of the flights table.
+
+```{r}
+options(rdflib_storage = "BDB")
+```
+
+
+## Tidyverse Style
+
+Operations in `dplyr` on the `nyflights13` dataset are easy to write and fast to execute, (in memory or on disk):  
+
+```{r}
+df <- flights %>% 
+  left_join(airlines) %>%
+  left_join(planes, by="tailnum") %>% 
+  select(carrier, manufacturer, model)
+```
+
+
+
+Use a smaller dataset if we do not have a BDB backend: 
+
+```{r}
+if(!rdf_has_bdb()){
+flights <- flights %>% 
+  filter(distance > 2500) # try smaller dataset
+}
+```
+
+
+Keys, including foreign keys, must be represented as URIs and not literal strings.  
+
+```{r}
+as_uri <- function(x, base_uri = "x:") paste0("<", base_uri, x, ">")
+
+flights <- flights %>% 
+  mutate(tailnum = as_uri(tailnum),
+         carrier = as_uri(carrier)) -> flights
+```
+
+
+
+# RDF Serialization Strategies for large data.frames
+
+We consider a variety of strategies for actually importing `data.frames` into RDF:
+
+- **Via `rdf_add()`**: Iterate over each row/cell with calls to `rdf_add`
+- **Via JSON-LD**: Coerce the `data.frame` to JSON (via `jsonlite::toJSON(force=TRUE)`), and parse as JSON-LD
+- **Via write.table()**` we `tidyr::gather()` and then hack such that we can call `write.table` on a `data.frame` to get an `nquads` text file
+
+As we'll see, only the third solution has adequate performance here. `rdf_add()` requires an initializer call inside each `redland::addStatement`,
+which takes a considerable fraction of a second.  Multiply that by the number of cells in the `data.frame` and things do not scale.
+
+`jsonlite` can convert even the large `data.frame`s into JSON reasonably quickly. 
+`jsonld::jsonld_to_rdf()` is then also acceptably fast (despite being Javascript) at converting this to `nquads`,
+but unfortunately fails dramatically (i.e. `segfault`) when attempting to serialize the flights data.  (Recall we can only get into redland RDF model from JSON-LD via nquads).
+Perhaps that is due to some particular data in `flights` table, but it's not obvious.  
+Otherwise, this approach has lots to recommend it.  One nice feature about this approach is that it applies to almost 
+any R object (e.g. any list object), though some care should be taken with names and URIs, as always. Another nice
+feature is that it handles the basic data types automatically -- JSON already has types for logical, double, integer, 
+and string, and these will get automatically encoded with the datatype URIs by the built-in `jsonld_to_rdf` algorithm.
+
+
[…]
