### 2018-02-16 8eabf8a
timing notes
