### 2018-02-17 21e50ae
Much faster, cleaner parsing of SPARQL returns
Uses getResults() method from redland package internals.  This is way way faster for returning large numbers of results.  This also sidesteps the need to rectangularize query results and manually coerce types; readr instead can handle that for us (as well as one can duck type from strings).
