+++ inst/examples/profile_performance.Rmd
@@ -32,7 +32,9 @@ Operations in `dplyr` on the `nyflights13` dataset are easy to write and fast to
 df <- flights %>% 
   left_join(airlines) %>%
   left_join(planes, by="tailnum") %>% 
-  select(carrier, manufacturer, model)
+  select(carrier, name, manufacturer, model) %>% 
+  distinct()
+head(df)
 ```
 
 
@@ -52,9 +54,14 @@ Keys, including foreign keys, must be represented as URIs and not literal string
 ```{r}
 as_uri <- function(x, base_uri = "x:") paste0("<", base_uri, x, ">")
 
-flights <- flights %>% 
+uri_flights <- flights %>% 
   mutate(tailnum = as_uri(tailnum),
-         carrier = as_uri(carrier)) -> flights
+         carrier = as_uri(carrier))
+
+
+uri_planes <- planes %>% 
+  mutate(tailnum = as_uri(tailnum),
+         carrier = as_uri(carrier))
 ```
 
 
@@ -106,6 +113,8 @@ With tables that have only tens of thousands of cells (triples) this is pretty f
 x1 <- as_rdf(airlines, "carrier", "x:")
 x2 <- as_rdf(airports, "faa", "x:") ## a few funny chars, UTF8 issues?
 x3 <- as_rdf(planes,  "tailnum", "x:")
+
+x1 <- c(x1,x2,x3)
 ```
 
 
@@ -113,21 +122,41 @@ SPARQL queries on the resulting data are also pretty fast:
 
 ```{r}
 sparql <-
-  'SELECT  ?seats ?model ?tailnum
+  'SELECT  ?seats ?model ?manufacturer ?p ?tailnum
 WHERE {
  ?tailnum <x:seats>  ?seats .
  ?tailnum <x:model>  ?model .
+ ?tailnum ?p  ?o .
+ ?tailnum <x:manufacturer>  ?manufacturer .
 }'
 
-out <- rdf_query(x3, sparql)
+out <- rdf_query(x1, sparql)
+head(out)
+unique(out$p)
+```
+
+
+
+```{r}
+sparql <-
+  'SELECT   ?model
+WHERE {
+ ?tailnum <x:carrier> ?carrier .
+ ?tailnum <x:model>  ?model 
+}'
+
+out <- rdf_query(x1, sparql)
 head(out)
 ```
+
+
+
 
 Big table via poor-man's `nquads` 165 seconds if this is the full table:
 
 ```{r}
 system.time(
-    x4 <- as_rdf(flights, NULL, "x:")
+    x4 <- as_rdf(uri_flights, NULL, "x:")
 )
 ```
 
@@ -145,11 +174,11 @@ The json-ld approach just appears to crash, so we won't run that:
 We can join all of these:  
 
 ```{r}
-rdf <- c(x4,x1,x2,x3)
+rdf <- c(x4,x1)
 ```
 
 
-Separate queries: This proves very slow on the full data!
+Separate queries: This proves very slow on the full data! Would be much faster if we did not have to iterate over getNextResult but could parse all results as a document.  Hopefully this change is coming to `redland` R library soon!
   
 ```{r}
 sparql <-
@@ -160,7 +189,7 @@ WHERE {
   ?flight <x:carrier>  ?carrier 
 }'
 
-profvis::profvis(
+system.time(
 f1 <- rdf_query(rdf, sparql)
 )
 ```
@@ -193,7 +222,36 @@ WHERE {
 }'
 
[…]
