### 2018-02-17 86376d8
move ex notebook to notebook/
