+++ inst/notebook/data-lake.md
@@ -15,6 +15,14 @@ source(system.file("examples/as_rdf.R", package="rdflib"))
 options(rdflib_storage = "memory") 
 ```
 
+``` r
+# Use a smaller dataset if we do not have a BDB backend: 
+if(getOption("rdflib_storage") != "BDB"){
+flights <- flights %>% 
+  filter(distance > 2600) # try smaller dataset
+}
+```
+
 Tidyverse Style
 ---------------
 
@@ -91,7 +99,7 @@ df <- rdf_query(rdf, s)
 ```
 
     ##    user  system elapsed 
-    ##   0.096   0.004   0.100
+    ##   0.263   0.015   0.310
 
 ``` r
 head(df)
@@ -112,3 +120,92 @@ Note that in place of joins, we give more semantically meaningful statements abo
 This is reminiscent of the way in which these data are organized in the relational database tables to begin with: we find `deb_delay` in the `flights` table and `manufacturer` in the `planes` table. Good relational design encourages this, but to work with the data the user is often left having to do the required joins, which also creates tables where these semantics are less clear.
 
 Tabular formats can often be sloppy about what is a key and what is a literal value, and also whether a column with the same name in different tables means the same thing in both. Both of these things pose challenges for later use when joining data. RDF representation encourages greater discipline through the use of URIs (though we've run a bit roughshod over that with the cavilier use of `x:` here.)
+
+Non-tabular data
+----------------
+
+``` r
+f <- system.file("extdata/gh_repos.json", package="repurrrsive")
+gh_data <- jsonlite::read_json(f)
+```
+
+Tidyverse style
+---------------
+
+``` r
+gh_flat <- gh_data %>% flatten()  # abandon nested structure and hope we didn't need it
+
+gh_tibble <- tibble(
+  name =     gh_flat %>% map_chr("name"),
+  issues =   gh_flat %>% map_int("open_issues_count"),
+  wiki =     gh_flat %>% map_lgl("has_wiki"),
+  homepage = gh_flat %>% map_chr("homepage", .default = ""),
+  owner =    gh_flat %>% map_chr(c("owner", "login"))
+)
+
+gh_tibble %>% arrange(name) %>% head()
+```
+
+    ## # A tibble: 6 x 5
+    ##   name                         issues wiki  homepage owner  
+    ##   <chr>                         <int> <lgl> <chr>    <chr>  
+    ## 1 2013-11_sfu                       1 T     ""       jennybc
+    ## 2 2014-01-27-miami                  4 T     ""       jennybc
+    ## 3 2014-05-12-ubc                    1 T     ""       jennybc
+    ## 4 2015_Coartic                      0 F     ""       jennybc
+    ## 5 2015-02-23_bryan-fields-talk      0 T     ""       jennybc
+    ## 6 2015-06-28_r-summit-talk          0 T     ""       jennybc
+
+RDF on non-tabular data
+-----------------------
+
+``` r
+gh_rdf <- as_rdf.list(gh_data, '"@vocab": "gh:"')
+```
+
+We can add it to the lake just for kicks:
+
+``` r
+system.time(
+rdf <- c(rdf, gh_rdf)
+)
+```
+
+    ##    user  system elapsed 
+    ##   2.954   0.060   3.217
+
+And we can query it back out of the lake
+
+``` r
+s <- 
+  'SELECT ?name ?issues ?wiki ?homepage ?owner
+WHERE {
+?repo <gh:homepage>  ?homepage .
+?repo <gh:has_wiki> ?wiki .
+?repo <gh:open_issues_count> ?issues .
+?repo <gh:name> ?name .
+?repo <gh:owner> ?owner_id .
+?owner_id <gh:login>  ?owner 
+}'
+
+system.time(
[…]
