+++ R/rdf_parse.R
@@ -29,22 +29,26 @@ rdf_parse <- function(doc,
                       ...){
   format <- match.arg(format)
   
+  ## if we get a string as input, we'll store it in tmp file here
+  ## which we can later be sure to clean up.
+  tmp_string <- tempfile()
+  ## if we get json-ld, we'll need a temp location to serialize that too:
+  tmp_json <- tempfile()
+  
   # convert string input or url to local file
-  doc <- text_or_url_to_doc(doc)
+  doc <- text_or_url_to_doc(doc, tmp_string)
   
   ## redlands doesn't support jsonld. So rewrite as nquads using jsonld package
   ## We use tmp to avoid altering input doc, since parsing a local file should
   ## be a read-only task!
   if(format == "jsonld"){
-    tmp <- tempfile()
-    #tmp <- add_base_uri(doc, tmp)
     x <- jsonld::jsonld_to_rdf(doc, 
                                options = 
            list(base = getOption("rdflib_base_uri", "localhost://"),
                 format = "application/nquads"))
-    writeLines(x, tmp)
+    writeLines(x, tmp_json)
     format <- "nquads"
-    doc <- tmp
+    doc <- tmp_json
   }
   
   if(is.null(rdf)){
@@ -54,8 +58,12 @@ rdf_parse <- function(doc,
   mimetype <- unname(rdf_mimetypes[format])
   parser <- new("Parser", rdf$world, name = format, mimeType = mimetype)
   redland::parseFileIntoModel(parser, rdf$world, doc, rdf$model)
-  redland::freeParser(parser)
   
+  redland::freeParser(parser)
+  unlink(tmp_string)
+  unlink(tmp_json)  
+  
+  ## return rdf object (pointer)
   rdf
 }
 
