### 2018-02-19 06170d4
make 'data-lake.Rmd' into vignette
