### 2018-02-19 cboettig/rdflib#13 @amoeba
Feedback on new rdf intro vignette
Hey Carl, I think this is fantastic. You've hit the nail on the head.

I thought I'd drop my feedback here rather than over on https://github.com/ropensci/onboarding/issues/169. I took a read-through for content and was very impressed with the current state of it and also with the sections you've penciled out for future work. I made a variety of minor notes as a list (immediately below) and made some high level comments in sub-sections below.

- I like the human tone the intro starts out on and I think the style is maintained throughout which was nice.
- Starting with the `tidyr::gather` equivalence is fantastic because I think it is a good bridge for users.
- You continually draw equivalencies for the user to technologies they are likely much more familiar with which is great.
- I really like Table 1!
- I really liked how you walked the user through tidying the `iris` data, using local IDs, the problem that arises, and how an RDF mindest would solve them.
- In the section "Subject URIs", when you start moving into namespaces, you offer up a namespace of "http://example.com/iris#" in order to world-wide-ify your IDs in the iris dataset. I think this might be a place you could start to lose users. For example, what is example.com/iris, why does this thing that looks like a URL not resolve to anything, and why would using broken URLs in my data be useful? I don't have any of the answers to these questions and I constantly struggle with and argue about this in my own work. The W3C intros use examples of Bob and Alice which, I think, make use of resolvable URIs. Do you think this stuff matters at this point in the intro?
- It's great that you show the example SPARQL query to pull the iris data into a tabular form
  but I think a better example to show graph queries (and thus graph thinking) might be useful.
  It kinda looks like you've put in a placeholder for this to happen already so 👍 I'd be happy to work on this part with the more graph-like tidyverse datasets.
- What do you think about renaming `rdf_serialize` to `rdf_write` to match the `{xml,html,etc}_write/_read` pattern in the tidyverse?
- What do you think about making `rdf_serialize` print to stdout by default? I realize `print.rdf` and `format.rdf` effectively do this, combined with the `rdf_print_format` option. This would be like how `writeLines` uses `con = stdout()` as a default.
- You use code like `rdf <- rdf()` which was a bit confusing as I didn't know that, somehow, R lets you do this. Do you think this could be a bit confusing?

### `rdf_add` API

I noticed this:

```r
rdf <- rdf()
rdf %>% rdf_add("",
                "iris:Sepal.Length",
                object = 5.1)
rdf
```

What do you think about adding support for first-class namespaces the way librdf does:

```python
n = Namespace("http://example.org/people/")
n.bob # = rdflib.term.URIRef(u'http://example.org/people/bob')
```

In R, this'd look similar, but you'd write like `n("bob")` instead:

```r
iris <- rdf_ns("http://example.com/iris#")

rdf <- rdf()
rdf %>% rdf_add("",
                iris("Sepal.Length"),
                object = 5.1)
rdf
```

Or maybe some more syntax-fu could be used, such as pipes or some other syntax magrittr provides. Maybe list assignment (`iris["obs1']`) or `/`(`iris / "obs1"`) can be overridden?

### The power (or utility?) of graph query langs

I think you do a nice job of showing graph queries, but when I've seen graph query languages (specifically GraphQL) demoed, the "why" is usually that, in graph query languages, we describe the shape of the result, rather than the instructing the database engine how to generate the result. SPARQL is very much one of these graph query languages and I think going into this could be useful. I wonder as I write this if there GraphQL is being adopted in R much yet.

### SPARQL, inference, and reasoners

A big problem with the RDF world is getting all of OWL's features to work.
When people create data, they often use OWL to create their own domain-specific ontology, rathern than making use of a pre-existing one. Many people would say this is a good thing (and many others would groan), because then they can map their ontology onto other ontologies and the linked data graph can grow. For example, if they're publishing their data with an additonal triple stating that their class `theirontology:Dataset`, is equivalent to, say, `Schema.org/Dataset`. Then, any query asking for a `Schema.org/Datset` would also return any `theirontology:Dataset`'s, in addition to any `Schema.org/Dataset`'s. The rub is that this requires an OWL reasoner to be running behind the SPARQL query engine, which I'm not sure is happening everywhere we run SPARQL, such as with redlands.

If this could me made to work for users of `rdflib` I think that would be a really useful feature. Perhaps this means helping the user get a triple store set up from within the package and use the RStudio Connection API to make it all fit within the RStudio ecosystem.

### More complex uses of `rdf_add`

`rdf_add` works great for adding single statements and can obviously be used to create as complex a data model as RDF can support. But could there be an easier way to create complex data models? Say I want to say that I have a husky, and want to use a URI for myself but not for my dog:

```r
rdf_data <- rdf()

# Manual blank nodes, probably already works?
rdf_data %>%
  rdf_add("http://example.org/Bryce",
          "pets:hasDog",
           "_:dog") %>%
  rdf_add("_:dog",
          "pets:isBreed",
           "Alaskan Husky")

# Nested `rdf_add` statements?
rdf_data %>%
  rdf_add("http://example.org/Bryce",
          "pets:hasDog",
           rdf_add(predicate = "pets:isBreed",
                   object = "Alaskan Husky"))

```

Or if I want to quickly add multiple statements about the same subject, like:

```r
rdf %>%
  rdf_add("ex:Me", c("ex:isSpecies", "ex:hasFirstName"), c("Human", "Bryce"))
```

Is the best way of inputting this type of stuff just a `data.frame`?
