+++ vignettes/data-lake.Rmd
@@ -1,0 +1,229 @@
+---
+title: The Data Lake and Schema On Read
+author: Carl Boettiger
+date: "`r Sys.Date()`"
+output: 
+  rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{The Data Lake and Schema On Read}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+  
+---
+
+
+```{r include = FALSE}
+knitr::opts_chunk$set(message=FALSE, warning = FALSE)
+```
+
+
+A provocative recent analogy (e.g. [Archer, 2017](https://www.w3.org/blog/2017/04/dive-into-the-semantic-data-lake/)) for thinking about RDF is that of the Data Lake.  Whereas adding new data to a traditional relational database frequently involves laborious wrangling into the an existing rigid data *schema* of tables and columns, RDF can enable a far more simple approach of just 'tossing everything into the data lake.' While this may sound like a recipe for a terrible mess down the road, RDF gives the promise of *schema-on-read*: rather than dictating the shape of your data when first adding it to the database (i.e. schema-on-write), a SPARQL query auto-magically reaches into the lake and retrieves all the relevant data, returning it in exactly the schema you ask for; that is, as a nice single table containing only the requested columns.  
+
+This can serve as a very effective means of data integration (provided a reasonably consistent and diligent use of URIs in identifying subjects and properties (predicates)), since just about any data can be added to the lake without worrying about whether it comes in a schema that matches the existing architecture of the database.  It is this flexibility not to have to define your database schema at the start that is the primary strength of the RDF approach. 
+
+A key advantage of this approach is that it is equally as easy to extract your desired data.frame from non-rectangular data as it is from tables or relational database systems. While the previous vignette focused on simple examples including data from single `data.frames` or small JSON files, this vignette showcases the Data Lake approach on a more complex relational database organized over several tables (`nyflights13` data, used to teach `joins` and other relational data in [Grolemund & Wickham, 2017](http://r4ds.had.co.nz)), and also in a large non-tabular file returned from the GitHub API (e.g. as used in [Bryan & Wickham, 2017 ](https://dcl-2017-04.github.io/curriculum/rectangling.html) to teach data rectangling.) 
+
+Load the necessary libraries to get started: 
+
+```{r libraries}
+## Data 
+library(nycflights13)
+library(repurrrsive)
+
+## for comparison approaches
+library(tidyverse)
+library(jsonlite)
+
+## Our focal package:
+library(rdflib)
+## experimental functions for rdflib package
+source(system.file("examples/as_rdf.R", package="rdflib"))
+```
+
+Configure RDF storage:
+
+```{r options, include = TRUE}
+#options(rdflib_storage = "BDB") 
+options(rdflib_storage = "memory") 
+
+# Use a smaller dataset if we do not have a BDB backend: 
+if(getOption("rdflib_storage") != "BDB"){
+flights <- flights %>% 
+  filter(distance > 2800) # try smaller dataset
+}
+```
+
+# Relational data
+
+
+## The `tidyverse` approach
+
+`tidyverse` operations are incredibly effective for working with relational data.  These `dplyr` on the `nyflights13` dataset are easy to write and fast to execute:
+
+```{r tidyverse}
+df <- flights %>% 
+  left_join(airlines) %>%
+  left_join(planes, by="tailnum") %>% 
+  select(carrier, name, manufacturer, model) %>% 
+  distinct()
+head(df)
+```
+ 
[…]
