+++ DESCRIPTION
@@ -37,15 +37,16 @@ Suggests: magrittr,
     testthat,
     knitr,
     rmarkdown,
-    jsonlite,
-    httr,
-    xml2,
     jqr,
-    lubridate,
     DT,
     tidyverse,
     dplyr,
     tidyr,
     tibble,
+    purrr,
+    lubridate,
+    httr,
+    xml2,
+    jsonlite,
     repurrrsive
 VignetteBuilder: knitr
