### 2018-02-19 ropensci/rdflib#13 @cboettig
Hey @amoeba , Thanks much, this is great and very helpful.  I think it would be very nice to flush these issues out further both in the package functions & vignettes better, and then sketch up a general-purpose introduction to RDF and related packages for R Journal with you and @annakrystalli.

A few quick replies while I read through this:

> you offer up a namespace of "http://example.com/iris#" in order to world-wide-ify your IDs in the iris dataset. I think this might be a place you could start to lose users.

Yes, I agree entirely; I haven't quite figured out how to introduce the whole concept of namespaces; but it would be much better if I could avoid them creeping in at all until I address the concept properly.  Maybe the issue of resolving vs non-resolving URIs should just be discussed at this stage?

> renaming `rdf_serialize` to `rdf_write`?

Yeah, this might be a good idea; the one thing that gives me pause is that write methods are serialization-specific: e.g. `readr` provides `write_csv`, `write_tsv`, `write_delim` for each possible serialization of the same `data.frame` structure.    Should we introduce `write_nquads`, `write_rdfxml`, `write_turtle` etc to match this pattern?   (I think getting the user comfortable with thinking about RDF as an abstraction with multiple serializations, rather than as something with a single `write_rdf`, is important to communicate in the API design here, but not sure which is best).

> rdf_serialize print to stdout by default

Sure, sounds reasonable.  Related, but the current `print` methods don't work with `head()` and don't stop printing, so they have issues (see #14).

> you use code like `rdf <- rdf()` ...

Ha, I blame you for that!  (You suggested the argument to `rdf_query`, `rdf_serialize` be called `rdf` instead of `x`, and so it made senes to me to use the same variable as the argument).  As you noticed,  this is fine (package function names belong to the package namespace / environment, not the global environment), but probably isn't a good practice.  Other choices for generic examples?  `my_rdf <- rdf()`?  `x <- rdf()`?

> ... better example to show graph queries ...

Definitely; I have a sketch of this now here: https://github.com/cboettig/rdflib/blob/patch-performance/vignettes/data-lake.Rmd

Again feedback would be great.  This vignette tries to focus on the more complex graph query examples and applications, whereas the `rdfintro.Rmd` is really more about figuring out how to introduce basic RDF concepts to the typical R user.  I think the "Data Lake" and "Schema on Read" metaphors are useful here, though these examples could definitely be improved to be made more interesting and show more complex cases (in particular, really should have an example of combining related but previously un-integrated data).

> complex uses of `rdf_add()`

Great idea.  Re nesting, I think the more natural "R" way is just to use lists!   Maybe something like:

```r
  list("@id" = "http://example.org/Bryce",
         "pets:hasDog" = list("pets:isBreed" = "Alaskan Husky")) %>%
as_rdf()
```

(this works with the experimental `as_rdf()` function, routing through JSON-LD of course...

gotta run, but more thoughts later
