### 2018-02-19 ropensci/rdflib#13 @cboettig
A few more quick thoughts:

Re:

```
# Manual blank nodes, probably already works?
rdf_data %>%
  rdf_add("http://example.org/Bryce",
          "pets:hasDog",
           "_:dog") %>%
  rdf_add("_:dog",
          "pets:isBreed",
           "Alaskan Husky")
```

Yup, this works, and in fact you can omit the `_:` and just do "dog" and you still get a blank node.  (Redland doesn't support `literal` subjects)

> namespaces for `rdf_add()`

This is a great idea. Though actually I think the whole namespace prefix thing isn't an optimal API -- it's familiar of course to anyone with an XML background but I think a stumbling block for most users.  I think what JSON-LD did with `@context` would actually be more intuitive (though I could be talked out of that).  That is:

-  you can define a default `@vocab` and then all your properties (predicates and URI-type objects) default to that namespace without a prefix.
- you can a define a `@base` (same as most RDF serializations?) which becomes the base URI for any blank node (e.g. subject with a string value instead of absolute URI).
- you can associate any individual term with a URI, e.g. `"name": "http://schema.org/name"`.

The latter is most interesting because on the face of it this looks more cumbersome; if you're seeped in RDF or XML we're more used to the notion that all terms come from specific vocabularies and we'd rather be explicit about it, saying this is a `dc:title` and that is a `schema:name`.  But for most users / programming / developer use it's much more natural to work with bare names (not least because in R you can say `list(name = "bob")` but need quotes to say `list("schema:name" = "bob")`....

So I think the namespace implementation should emulate the ability to essentially define a context (using an R list), rather than a prefix per se.  Of course you can define prefixes in a context as well (and also define types, e.g. a URI vs literal), which is super nice.  So this needs some work to implement but I'd like to support bare names with a full mapping rather than only support prefixes.

> GraphQL ...

Scott has https://github.com/ropensci/ghql which wraps Jeroen's bindings at https://github.com/ropensci/graphql .  I think GraphQL is really elegant; but the backend setup to create a graphQL database looks complex to me: basically asks you to create your own ontology of types and for some reason does not reuse and RDF notions.  But haven't looked closely enough at it yet to honestly judge...

> OWL ....

The vignette should definitely mention this somewhere to define Ontology, OWL, and the role of a reasoner.  But I'm still warming up to OWL, with the following objections:

1. As you know, we don't have access to any reasoner without a Java wrapper.  Would love a good reasoner as a  C library ...

2. For a most use cases, the user won't have access to a good ontology.  (Even finding a good namespace is often hard).  Of course it would still make sense to support ontologies when they are available.   An ideal approach should probably also support easy construction of common OWL clauses that a user could write on the fly to add reasoning, but:

3. I'm still skeptical of the utility of general-purpose reasoning.  (probably my own ignorance, I could definitely do a 180 on this).   My current thinking is that R users (or me at any rate) would find it easier to introduce reasoning on triples by writing specific functions that operate on their triplestore than by writing out OWL logic.  For instance, the phenology ontology gives the example of inferring the triple "hasFlowers  = true" from a statement of `flowerCount = 5` -- this seems to me the kind of thing that would be more natural to do as a function call on a database, and hard to imagine an ontology successfully anticipating all such possible deductions ahead of time.  More generally, any `mutate` operation in `dplyr` is essentially a logical inference from existing triplestore that adds more triples, right?  OWL feels to me like it is essentially a subset of such functions?

Other examples might include inferring membership of a containing class; e.g. identify all frogs from a species list, also seem more natural as functions or table joins of species ids to a table of the ranks.  Making the user specify their inferences in  their own functions also seems like the best way to avoid the dangers of poorly specified OWL relationships leading to nonsensical conclusions.  But I could be entirely wrong on this -- I'd love to have a compelling example or two of OWL use that we could compare against doing the same inference with functions.

I also suspect that a philosophical difference between the "data science" mindset of today and the reasoning behind the design of OWL (and also SQL as it applies to being a user interface) is the assumption that the user does not have a convenient general-purpose programming language at their disposal, and thus doesn't have the option of just writing on-the-fly functions to perform inference...`dplyr` (and `pandas`) brought the speed and abstractions of SQL to users inside of a general-purpose programming environment rather than a stand-alone RDBM client, which I think has proven very powerful.  I think there are similarly powerful abstractions in RDF which are really just an extension of concepts already in relational database design we can leverage.
