### 2018-02-20 a49fb6a
rdf_add can handle NA as a blank node
