### 2018-02-20 e8f2927
c() method use turtle to save disk space
