### 2018-02-20 887294a
parser and serializer will guess format based on file extension, closes #4
serializer also sets explicit base option

serializer defaults to print to character string if doc is NULL.
