### 2018-02-20 39074cf
cleaning up as_rdf methods
methods take vocab, base, and key
