+++ inst/examples/as_rdf.R
@@ -1,4 +1,45 @@
-as_rdf <- function(df, key = NULL, base_uri = NULL, loc = NULL) UseMethod("as_rdf")
+
+#' As RDF
+#' 
+#' Coerce an object into RDF
+#' @param x an object to coerce into RDF (list, list-like, or data.frame)
+#' @param rdf An existing rdf object, (by default a new object will be initialized)
+#' @param vocab A default vocabulary (URI prefix) to assume for all predicates
+#' @param base A base URI to assume for blank subject nodes
+#' @param context: a named list mapping any string to a URI
+#' @param ... additional arguments, specific to certain types
+#' 
+as_rdf <- function(x, 
+                   rdf = rdf(),
+                   vocab = NULL, 
+                   base = getOption("rdflib_base_uri"), 
+                   context = NULL, 
+                   ...) UseMethod("as_rdf")
+
+as_rdf.list <- function(x, 
+                        vocab = NULL, 
+                        base = getOption("rdflib_base_uri"), 
+                        context = NULL){
+  
+  ## unbox length-1 lists so we can apply a context successfully
+  if(is(x, "list") && length(x) == 1) x <- x[[1]]
+  
+  json <- jsonlite::toJSON(x, pretty = TRUE, auto_unbox = TRUE, force = TRUE)
+  jsonld_context <- json_context(vocab, base, context)
+  json2 <- paste0('{\n"@context":', jsonld_context, ',\n',  '"@graph": ', json,  '}')
+  rdf <- rdflib::rdf_parse(json2, "jsonld", rdf = rdf)
+  rdf
+}
+# helper function (identical to plyr::compact)
+compact <- function (l) Filter(Negate(is.null), l)
+json_context <- function(vocab = NULL, base = getOption("rdflib_base_uri"), context = NULL){
+  jsonlite::toJSON(
+    compact(c(list("@vocab" = vocab,
+                   "@base" = base),
+              context)), auto_unbox = TRUE)
+}
+
+
 
 # test2 <- tibble(
 #   age = 5L,
@@ -7,42 +48,61 @@ as_rdf <- function(df, key = NULL, base_uri = NULL, loc = NULL) UseMethod("as_rd
 #   address = "x:address",
 #   knows = NA
 # )
-# as_rdf.data.frame(test2, NULL, "x:", "test2.nquads")
+# as_rdf.data.frame(test2, vocab = "x:", key = "nquads")
 
 ## tidy data to rdf
-as_rdf.data.frame <- function(df, key = NULL, base_uri = NULL, loc = tempfile()){
-  
-  x <- df
-  if(is.null(key)){
-    x <- tibble::rowid_to_column(x, "subject")
-  } else {
-    names(x)[names(x) == key] <- "subject"
-  }
-  
-  ## FIXME consider taking an already-gathered table to avoid dependency?
-  suppressWarnings(
-    x <- tidyr::gather(x, key = predicate, value = object, -subject)
-  )
-  
-  
+as_rdf.data.frame <- function(df,  
+                              rdf = rdf(), 
+                              vocab = NULL, 
+                              base = getOption("rdflib_base_uri"), 
+                              context = NULL, 
+                              key = NULL){
   
   ## gather looses col-classes, so pre-compute them (with base R)
   col_classes <- data.frame(datatype = 
                               vapply(df, 
                                      rdflib:::xs_class, 
-                                     character(1)),
-                            stringsAsFactors = FALSE)
[…]
