### 2018-02-20 481a329
datatype should not be assigned to blank nodes
