### 2018-02-20 5574938
use rdflib_base_uri throughout
