### 2018-02-20 3c3aceb
avoid c() by passing rdf arg
