### 2018-02-20 38598ea
option to reconnect to an existing database
