### 2018-02-20 7a02e04
indicate storage type in rdf() constructor instead
