+++ tests/testthat/test-as_rdf.R
@@ -1,0 +1,3 @@
+testthat::context("as_rdf")
+
+## tests coming, see vignettes
