+++ R/rdf_query.R
@@ -32,12 +32,11 @@
 #'
 rdf_query <- function(rdf, query, data.frame = TRUE, ...){
   queryObj <- new("Query", rdf$world, query, ...)
-  # ... defaults are: base_uri=NULL, query_language="sparql", query_uri=NULL)
+  
+  # ... defaults are: base_uri=NULL, query_language="sparql", query_uri=NULL
+  
   queryResult <- redland::executeQuery(queryObj, rdf$model)
   out <- getResults(queryResult)
-  
-  #iter_getNextResult(queryResult)
-  # out <- rectangularize_query_results(out)
   redland::freeQueryResults(queryResult)
   redland::freeQuery(queryObj)
   
@@ -62,44 +61,3 @@ getResults <- function(queryResult, format = "csv", ...){
                   ...)
 }
 
-
-
-
-
-## Deprecate this strategy in favor of direct result
-## also means we can deprecate type_by_datauri and other utilities
-iter_getNextResult <- function(queryResult){
-  out <- list()
-  result <- redland::getNextResult(queryResult)
-  out <- c(out, result)
-  while(!is.null(result)){
-    result <- redland::getNextResult(queryResult)
-    out <- c(out, result)
-  }
-  out
-}
-
-rectangularize_query_results <- function(out){
-  vars <- unique(names(out))
-  
-  X <- lapply(vars, function(v){ 
-    contents <- as.character(out[names(out) == v ])
-    values <- type_by_datauri(contents)
-    
-    ## use "character" if mixed type column
-    types <- vapply(values, function(x) class(x)[[1]], character(1))
-    u <- unique(types)
-    if(length(u) == 1){
-      values <- unlist(values)
-      if(u %in% c("Date", "POSIXct"))
-        class(values) <- unique(types) # Restore date class
-    } else {
-      values <- vapply(values, as.character, character(1))
-    }
-    values
-  })
-  
-  names(X) <- vars
-  ## Or we could use tibble::as_data_frame for list columns w/ mixed type..
-  as.data.frame(X, stringsAsFactors=FALSE)
-}
