+++ tests/testthat/test-as_rdf.R
@@ -1,3 +1,4 @@
 testthat::context("as_rdf")
 
 ## tests coming, see vignettes
+
