### 2018-02-20 ec81511
skip has_bdb on appveyor
