### 2018-02-20 01f59bb
Patch performance (#16)
* Much faster, cleaner parsing of SPARQL returns

Uses getResults() method from redland package internals.  This is way way faster for returning large numbers of results.  This also sidesteps the need to rectangularize query results and manually coerce types; readr instead can handle that for us (as well as one can duck type from strings).

* suggest nycflights13 data

* rdf_add can handle NA as a blank node

* c() method use turtle to save disk space

* parser and serializer will guess format
based on file extension, closes #4

* serializer also sets explicit base opti
[…]
