+++ NAMESPACE
@@ -19,7 +19,6 @@ importClassesFrom(redland,Storage)
 importClassesFrom(redland,World)
 importFrom(jsonld,jsonld_compact)
 importFrom(jsonld,jsonld_to_rdf)
-importFrom(methods,as)
 importFrom(methods,new)
 importFrom(readr,read_csv)
 importFrom(stringi,stri_unescape_unicode)
