### 2018-02-20 #169 @lmullen
Hi @cboettig. Thanks for the thoroughness of your response to the reviews. At this point I don't see any reason to delay accepting the package into rOpenSci. Of course it looks like you are still figuring out the final form of a few things, especially in the new vignette, so it will be your call when to submit to CRAN.

@karthik I don't have access to rOpenSci admin accounts, so could you please begin the process of moving this package into the rOpenSci organization?

After that happens, @cboettig, could you please do the following?

- [ ] Add the rOpenSci footer
- [ ] Check that all CI works in the new rOpenSci repository

Once the repository is moved I will close this issue.
