### 2018-02-20 ropensci/rdflib#10 @cboettig
All resolved and approved.
