### 2018-02-20 ropensci/rdflib#9 @cboettig
These issues are now resolved and approved.
