### 2018-02-21 94f673b
Update README, bump version, add badge, add CITATION
