### 2018-02-21 f8bbcd8
build docs for ropensci
