### 2018-02-21 #169 @cboettig
Thanks @lmullen and @karthik!  I've added the ropensci footer and migrated the repo to `ropensci`.  🚀 .

CI seems to be working (if I  recall correctly I shouldn't migrate the appveyor since it only links to individual accounts?)

I assume there's nothing I need to do to update the onboarding badge, that happens automatically via tags on this issue, right?

I'll leave it to you editors to close out this thread when ready.
