### 2018-02-21 #169 @karthik
Thanks @cboettig & @lmullen!!
